pyproject.toml
[project]
name = "sqlagg"
authors = [{name = "Dimagi", email = "dev@dimagi.com"}]
license = {file = "LICENSE"}
readme = {file = "README.md", content-type = "text/markdown"}
dynamic = ["version", "description"]
requires-python = ">= 3.9"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    'SQLAlchemy>=1.0.9,<1.4',
]

[project.urls]
Home = "https://github.com/dimagi/sql-agg"

[project.optional-dependencies]
test = [
    "SQLAlchemy-Fixtures>=0.1.5",
    "pytest",
    "pytest-unmagic",
    "psycopg2-binary",
]

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
